+++ packages/block-editor/src/components/index.js
@@ -76,13 +76,15 @@ export {
 	__unstableRichTextInputEvent,
 } from './rich-text';
 export { default as ToolSelector } from './tool-selector';
-export { default as __experimentalUnitControl } from './unit-control';
+export {
+	default as __experimentalUnitControl,
+	useCustomUnits,
+} from './unit-control';
 export { default as URLInput } from './url-input';
 export { default as URLInputButton } from './url-input/button';
 export { default as URLPopover } from './url-popover';
 export { __experimentalImageURLInputUI } from './url-popover/image-url-input-ui';
 export { default as withColorContext } from './color-palette/with-color-context';
-export { useCustomUnits } from './use-custom-units';
 
 /*
  * Content Related Components
